+++ storeapp/src/main/java/com/localweb/storeapp/entity/Role.java
@@ -8,7 +8,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private int id;
+    private long id;
 
     @Column(name = "name")
     private String name;
@@ -16,7 +16,7 @@ public class Role {
     public Role() {
     }
 
-    public Role(int id) {
+    public Role(long id) {
         this.id = id;
     }
 
@@ -24,16 +24,16 @@ public class Role {
         this.name = name;
     }
 
-    public Role(int id, String name) {
+    public Role(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
